refactor: move text editor modules
