+++ BlogposterCMS/public/assets/scss/components/_color-picker.scss
@@ -34,6 +34,15 @@
     color: #555;
   }
 
+  .color-input {
+    width: 60px;
+    padding: 2px 4px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    font-size: 12px;
+    outline: none;
+  }
+
   .color-section-label {
     margin-right: 4px;
     font-size: 12px;
